+++ packages/core/story-structure.js
@@ -46,12 +46,10 @@ const s = [
             n: 'Datagrid',
             s: [
               'c/Datagrid',
-              'c/Datagrid/Extensions/Header',
               'c/Datagrid/Extensions/RowHeightSettings',
               'c/Datagrid/Extensions/RowActionButtons',
               'c/Datagrid/Extensions/ExpandableRow',
               'c/Datagrid/Extensions/NestedRows',
-              'c/Datagrid/Extensions/ColumnAlignment',
               'c/Datagrid/Extensions/ClickableRow',
               'c/Datagrid/Extensions/EditableCell',
               'c/Datagrid/Extensions/ColumnCustomization',
@@ -124,6 +122,7 @@ const s = [
           'c/ButtonMenu',
           'c/ButtonSetWithOverflow',
           'c/CancelableTextEdit',
+          'c/Carousel',
           'c/ComboButton',
           'c/ExampleComponent',
           'c/FilterSummary',
@@ -135,9 +134,25 @@ const s = [
       },
       {
         n: 'Novice to pro',
-        s: ['c/Checklist', 'c/InlineTip'],
+        s: [
+          'c/Checklist',
+          'c/Guidebanner',
+          'c/InlineTip',
+          'c/NonLinearReading',
+        ],
       },
     ],
+  },
+  {
+    n: 'Community',
+    s: [
+      {
+        n: 'Components',
+        s: [
+          { n: 'Example', s: ['c/ExampleCommunityComponent'] }
+        ]
+      }
+    ]
   },
   {
     n: 'Carbon',
@@ -157,14 +172,14 @@ const getSectionOrder = (sArray) =>
     .map((entry) =>
       typeof entry === 'string'
         ? // if the entry is a string, return it (without the prefix)
-          getEntryDisplayName(entry)
+        getEntryDisplayName(entry)
         : // if the entry is another structure, return its name, but first get
-          // the fully materialized paths it contains and add the entry name
-          // to the front of each
-          [
-            getSectionOrder(entry.s).map((path) => `${entry.n}/${path}`),
-            entry.n,
-          ]
+        // the fully materialized paths it contains and add the entry name
+        // to the front of each
+        [
+          getSectionOrder(entry.s).map((path) => `${entry.n}/${path}`),
+          entry.n,
+        ]
     )
     .flat(Infinity);
 
@@ -197,13 +212,13 @@ const getPath = (s, componentAndPrefix, componentName) =>
       found ||
       (typeof next === 'string'
         ? // if this entry is a string, and it matches the component name
-          // we're looking for, return it as an array, else return null
-          getEntryPrefixAndComponentName(next) === componentAndPrefix
+        // we're looking for, return it as an array, else return null
+        getEntryPrefixAndComponentName(next) === componentAndPrefix
           ? [getEntryDisplayName(next)]
           : null
         : // if this entry is another structure, find the materialized path
-          // into it and prepend its name if found and return null otherwise
-          prepend(next.n, getPath(next.s, componentAndPrefix, componentName))),
+        // into it and prepend its name if found and return null otherwise
+        prepend(next.n, getPath(next.s, componentAndPrefix, componentName))),
     null
   );