+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridRow.js
@@ -24,7 +24,15 @@ const rowHeights = {
 
 // eslint-disable-next-line react/prop-types
 const DatagridRow = (datagridState) => {
-  const { row, rowSize, withNestedRows, prepareRow, key } = datagridState;
+  const {
+    row,
+    rowSize,
+    withNestedRows,
+    prepareRow,
+    key,
+    tableId,
+    withExpandedRows,
+  } = datagridState;
 
   const getVisibleNestedRowCount = ({ isExpanded, subRows }) => {
     let size = 0;
@@ -67,7 +75,7 @@ const DatagridRow = (datagridState) => {
 
   const focusRemover = () => {
     const elements = document.querySelectorAll(
-      `.${blockClass}__carbon-row-expanded`
+      `#${tableId} .${blockClass}__carbon-row-expanded`
     );
     elements.forEach((el) => {
       el.classList.remove(`${blockClass}__carbon-row-expanded-hover-active`);
@@ -107,21 +115,35 @@ const DatagridRow = (datagridState) => {
     [`${carbon.prefix}--data-table--selected`]: row.isSelected,
   });
 
+  const setAdditionalRowProps = () => {
+    if (withNestedRows || withExpandedRows) {
+      return {
+        'data-nested-row-id': row.id,
+      };
+    }
+    return {};
+  };
+
+  const { role, ...rowProps } = row.getRowProps();
+
   return (
     <React.Fragment key={key}>
       <TableRow
         className={rowClassNames}
-        {...row.getRowProps()}
+        {...rowProps}
         key={row.id}
         onMouseEnter={hoverHandler}
         onMouseLeave={handleMouseLeave}
         onFocus={hoverHandler}
         onBlur={focusRemover}
         onKeyUp={handleOnKeyUp}
+        {...setAdditionalRowProps()}
+        // avoid unnecessary role assignment to rows
+        {...(role === 'row' && { role })}
       >
         {row.cells.map((cell, index) => {
           const cellProps = cell.getCellProps();
-          const { children, ...restProps } = cellProps;
+          const { children, role, ...restProps } = cellProps;
           const content = children || (
             <>
               {!row.isSkeleton && cell.render('Cell')}
@@ -140,6 +162,8 @@ const DatagridRow = (datagridState) => {
               })}
               {...restProps}
               key={cell.column.id}
+              // avoid unnecessary role assignment to cells
+              {...(role === 'cell' && { role })}
             >
               {content}
             </TableCell>