+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridToolbar.js
@@ -103,6 +103,7 @@ const DatagridBatchActionsToolbar = (datagridState, width, ref) => {
             [`${menuClass}-icon-only`]: width <= minWidthBeforeOverflowIcon,
           },
         ])}
+        tabIndex={totalSelected > 0 ? 0 : -1}
       >
         {toolbarBatchActions?.map((batchAction, index) => {
           const hidden = index < 2 && !displayAllInMenu;
@@ -161,6 +162,7 @@ const DatagridBatchActionsToolbar = (datagridState, width, ref) => {
                 renderIcon={batchAction.renderIcon}
                 onClick={(event) => onClickHandler(event, batchAction)}
                 iconDescription={batchAction.label}
+                tabIndex={totalSelected > 0 ? 0 : -1}
               >
                 {batchAction.label}
               </TableBatchAction>
@@ -186,7 +188,7 @@ const DatagridToolbar = (datagridState) => {
       className={cx([blockClass, `${blockClass}--${getRowHeight}`])}
     >
       <TableToolbar>
-        {DatagridActions && DatagridActions(datagridState)}
+        {DatagridActions && <DatagridActions {...datagridState} />}
         {DatagridBatchActionsToolbar &&
           DatagridBatchActionsToolbar(datagridState, width, ref)}
       </TableToolbar>
@@ -194,7 +196,7 @@ const DatagridToolbar = (datagridState) => {
   ) : DatagridActions ? (
     <div className={blockClass}>
       <TableToolbar>
-        {DatagridActions && DatagridActions(datagridState)}
+        {DatagridActions && <DatagridActions {...datagridState} />}
         {DatagridBatchActions && DatagridBatchActions(datagridState)}
       </TableToolbar>
     </div>