+++ packages/ibm-products/src/components/InlineTip/InlineTip.stories.js
@@ -81,6 +81,7 @@ const defaultProps = {
     action(`Clicked the close button`)();
   },
   title: 'Use case-specific heading',
+  withLeftGutter: false,
 };
 
 const Template = (args) => {