+++ packages/ibm-products/src/components/NonLinearReading/NonLinearReading.docs-page.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { StoryDocsPage } from '../../global/js/utils/StoryDocsPage';
+
+import * as stories from './NonLinearReading.stories';
+
+const DocsPage = () => (
+  <StoryDocsPage
+    blocks={[
+      {
+        story: stories.singleLevel,
+      },
+      {
+        story: stories.multipleLevel,
+      },
+      {
+        story: stories.withGradientBackground,
+      },
+    ]}
+  />
+);
+
+export default DocsPage;