+++ packages/ibm-products/src/components/ProductiveCard/ProductiveCard.js
@@ -149,6 +149,11 @@ ProductiveCard.propTypes = {
    */
   secondaryButtonText: PropTypes.node,
   /**
+   * **Experimental:** For all cases a `Slug` component can be provided.
+   * Clickable tiles only accept a boolean value of true and display a hollow slug.
+   */
+  slug: PropTypes.oneOfType([PropTypes.node, PropTypes.bool]),
+  /**
    * Title that's displayed at the top of the card
    */
   title: PropTypes.oneOfType([